+++ src/Ormolu/Printer/Meat/Type.hs
@@ -25,7 +25,7 @@ module Ormolu.Printer.Meat.Type
   )
 where
 
-import Data.Choice (pattern With)
+import Control.Monad
 import GHC.Data.Strict qualified as Strict
 import GHC.Hs hiding (isPromoted)
 import GHC.Types.SourceText
@@ -116,10 +116,8 @@ p_hsType' multilineArgs = \case
     let opTree = BinaryOpBranches (tyOpTree x) op (tyOpTree y)
     p_tyOpTree
       (reassociateOpTree debug (Just . unLoc) modFixityMap opTree)
-  HsParTy _ t -> do
-    csSpans <-
-      fmap (flip RealSrcSpan Strict.Nothing . getLoc) <$> getEnclosingComments
-    switchLayout (locA t : csSpans) $
+  HsParTy _ t ->
+    switchLayoutWithEnclosingComments [locA t] $
       parens N (located t p_hsType)
   HsIParamTy _ n t -> sitcc $ do
     located n atom
@@ -136,7 +134,7 @@ p_hsType' multilineArgs = \case
     inci (located k p_hsType)
   HsSpliceTy _ splice -> p_hsUntypedSplice DollarSplice splice
   HsDocTy _ t str -> do
-    p_hsDoc Pipe (With #endNewline) str
+    p_hsDocInline Pipe str
     located t p_hsType
   HsExplicitListTy _ p xs -> do
     case p of
@@ -274,11 +272,14 @@ p_forallBndrs vis p tyvars =
 
 p_hsConDeclRecFields :: [LHsConDeclRecField GhcPs] -> R ()
 p_hsConDeclRecFields xs =
-  braces N $ sep commaDel (sitcc . located' p_hsConDeclRecField) xs
+  multiLineIfDocumented xs . braces N $ do
+    when (null xs) $
+      getEnclosingSpan >>= mapM_ (locatedEmpty . flip RealSrcSpan Strict.Nothing)
+    sep commaDel (sitcc . located' p_hsConDeclRecField) xs
 
 p_hsConDeclRecField :: HsConDeclRecField GhcPs -> R ()
 p_hsConDeclRecField HsConDeclRecField {..} = do
-  mapM_ (p_hsDoc Pipe (With #endNewline)) (cdf_doc cdrf_spec)
+  mapM_ (p_hsDocInline Pipe) (cdf_doc cdrf_spec)
   sitcc $
     sep
       commaDel
@@ -308,7 +309,7 @@ p_hsConDeclField CDF {..} = do
 
 p_hsConDeclFieldWithDoc :: HsConDeclField GhcPs -> R ()
 p_hsConDeclFieldWithDoc cdf = do
-  mapM_ (p_hsDoc Pipe (With #endNewline)) (cdf_doc cdf)
+  mapM_ (p_hsDocInline Pipe) (cdf_doc cdf)
   p_hsConDeclField cdf
 
 p_lhsTypeArg :: LHsTypeArg GhcPs -> R ()